+++ src/Faker/Extensions/IEnumerableExtensions.cs
@@ -1,7 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 
 namespace RimuTec.Faker.Extensions {
    internal static class IEnumerableExtensions {